Make tiptap more reusable by defining own classes
