+++ dc-cudami-admin/src/main/resources/templates/fragments/tiptap.html
@@ -62,38 +62,89 @@
   </th:block>
 
   <th:block th:fragment="editor-menu-short">
-    <div id="editorMenu" class="editorMenu border-1 p-1">
-      <div class="btn-group" role="group" aria-label="...">
-        <button type="button" class="btn btn-default" tabindex="-1" title="Undo" disabled="true" data-value="undo"
-                onclick="editor.chain().focus().undo().run()"><i class="fa fa-undo"></i></button>
+    <div class="btn-group" role="group" aria-label="...">
+      <button type="button" class="btn btn-default" tabindex="-1" title="Undo" disabled="true" data-value="undo"
+              onclick="editor.chain().focus().undo().run()"><i class="fa fa-undo"></i></button>
 
-        <button type="button" class="btn btn-default" tabindex="-1" title="Redo" disabled="true" data-value="redo"
-                onclick="editor.chain().focus().redo().run()"><i class="fa fa-redo"></i></button>
-      </div>
-      &nbsp;
-      <div class="btn-group" role="group" aria-label="...">
-        <button type="button" class="btn btn-default" tabindex="-1" title="Bold" data-value="bold"
-                onclick="editor.chain().focus().toggleBold().run()"><span class="fa fa-bold"></span></button>
+      <button type="button" class="btn btn-default" tabindex="-1" title="Redo" disabled="true" data-value="redo"
+              onclick="editor.chain().focus().redo().run()"><i class="fa fa-redo"></i></button>
+    </div>
+    &nbsp;
+    <div class="btn-group" role="group" aria-label="...">
+      <button type="button" class="btn btn-default" tabindex="-1" title="Bold" data-value="bold"
+              onclick="editor.chain().focus().toggleBold().run()"><span class="fa fa-bold"></span></button>
 
-        <button type="button" class="btn btn-default" tabindex="-1" title="Italic" data-value="italic"
-                onclick="editor.chain().focus().toggleItalic().run()"><span class="fa fa-italic"></span></button>
+      <button type="button" class="btn btn-default" tabindex="-1" title="Italic" data-value="italic"
+              onclick="editor.chain().focus().toggleItalic().run()"><span class="fa fa-italic"></span></button>
 
-        <button type="button" class="btn btn-default" tabindex="-1" title="Underline" data-value="underline"
-                onclick="editor.chain().focus().toggleUnderline().run()"><span class="fa fa-underline"></span></button>
+      <button type="button" class="btn btn-default" tabindex="-1" title="Underline" data-value="underline"
+              onclick="editor.chain().focus().toggleUnderline().run()"><span class="fa fa-underline"></span></button>
 
-        <button type="button" class="btn btn-default" tabindex="-1" title="Strike" data-value="strike"
-                onclick="editor.chain().focus().toggleStrike().run()"><span class="fa fa-strikethrough"></span></button>
+      <button type="button" class="btn btn-default" tabindex="-1" title="Strike" data-value="strike"
+              onclick="editor.chain().focus().toggleStrike().run()"><span class="fa fa-strikethrough"></span></button>
 
-        <button type="button" class="btn btn-default" tabindex="-1" title="Subscript" data-value="subscript"
-                onclick="editor.chain().focus().toggleSubscript().run()"><span class="fa fa-subscript"></span></button>
+      <button type="button" class="btn btn-default" tabindex="-1" title="Subscript" data-value="subscript"
+              onclick="editor.chain().focus().toggleSubscript().run()"><span class="fa fa-subscript"></span></button>
 
-        <button type="button" class="btn btn-default" tabindex="-1" title="Superscript" data-value="superscript"
-                onclick="editor.chain().focus().toggleSuperscript().run()"><span class="fa fa-superscript"></span></button>
+      <button type="button" class="btn btn-default" tabindex="-1" title="Superscript" data-value="superscript"
+              onclick="editor.chain().focus().toggleSuperscript().run()"><span class="fa fa-superscript"></span></button>
 
-        <button type="button" class="btn btn-default" tabindex="-1" title="Link" data-value="link"
-                onclick="editor.chain().focus().setLink({ href: 'http://www.google.de/'}).run()"><span class="fa fa-link"></span></button>
+      <button type="button" class="btn btn-default" tabindex="-1" title="Link" data-value="link"
+              data-target="#dialog-link" data-toggle="modal"><span class="fa fa-link"></span></button>
+      <!--        <button type="button" class="btn btn-default" tabindex="-1" title="Link" data-value="link"
+                      onclick="editor.chain().focus().setLink({href: 'http://www.google.de/'}).run()"><span class="fa fa-link"></span></button>-->
+    </div>
+  </th:block>
+
+  <th:block th:fragment="modal-dialog-editor-link">
+    <div class="modal fade" id="dialog-link">
+      <div class="modal-dialog">
+        <div class="modal-content">
+          <div class="modal-header">
+            <!-- TODO: title depending on add/edit link -->
+            <h4 class="modal-title">Add Link</h4>
+            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          <form class="">
+            <div class="modal-body">
+              <div class="form-group">
+                <label for="link-url-input">URL</label>
+                <input id="link-url-input" name="url" pattern="^(https?://|mailto:|/|#).*$" required="" type="text" class="form-control" value="">
+              </div>
+              <div class="form-group">
+                <label for="link-title-input">Tooltip</label>
+                <input id="link-title-input" name="title" type="text" class="form-control" value="">
+              </div>
+            </div>
+            <div class="modal-footer justify-content-between">
+              <button type="button" class="btn btn-default" data-dismiss="modal" th:text="#{btn.cancel}">Cancel</button>
+              <a class="btn btn-primary" id="confirm" th:text="#{btn.add}">Add</a>
+            </div>
+          </form>
+        </div>
       </div>
     </div>
+
+    <script type="text/javascript">
+      $("#dialog-link").on('click', '.btn-primary', function () {
+        let linkUrl = $("#dialog-link input[name='url']").val();
+        let linkTitle = $("#dialog-link input[name='title']").val();
+        /*
+         * let attrs={ title: 'mylink', href:'google.com' }
+         let schema=view.state.schema
+         let node=schema.text(attrs.title, [schema.marks.link.create(attrs)])
+         view.dispatch(view.state.tr.replaceSelectionWith(node, false))
+         */
+        if (linkTitle) {
+          editor.chain().focus().extendMarkRange("link").setLink({href: linkUrl, title: linkTitle}).run();
+        } else {
+          editor.chain().focus().extendMarkRange("link").setLink({href: linkUrl}).run();
+        }
+        $('#dialog-link').modal('hide');
+      });
+    </script>
   </th:block>
 </body>
 </html>
